March Update
Project now includes some music sourced from Open Game Art.  Attribution file should be made later.  Cleaned up Block Manager by moving the map generation code into it's own class file.  Worked a bit more on the AI system.
The walking Ai is nearly ready  to be implemented in a more proper system.
I'll be slowly working on more documentation as the game expands, but it's not exactly a top priority.

This project might end fairly soon.  I think I'll plan a few more systems but a few things make me feel as if Unity isn't the best engine for a game like this, and I'm not experienced enough in the backend of 2D graphics to optimize it properly to make it work better.